// packages/vendure-plugin-shopify-app-bridge/src/shopify/webhook/types.ts
export interface CustomerWebhookShopify {
  id: number;
  email: string;
  first_name: string;
  last_name: string;
  state: string;
  note: string;
  tags: string;
  currency: string;
  phone: string;
  addresses: Array<AddressWebhookShopify>;
  default_address: AddressWebhookShopify;
}

export interface AddressWebhookShopify {
  id: number;
  customer_id: number;
  first_name: string;
  last_name: string;
  company: string;
  address1: string;
  address2: string;
  city: string;
  province: string;
  country: string;
  zip: string;
  phone: string;
  name: string;
  province_code?: string;
  country_code: string;
  country_name: string;
  default: boolean;
}

export interface CustomerDeleteWebhookShopify {
  id: number;
}